% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fluid_summation.R
\name{calculate_fluid_summation}
\alias{calculate_fluid_summation}
\title{Calculate the fluids received and removed from the patient.}
\usage{
calculate_fluid_summation(temp_INS, temp_OUTS, temp_urine, temp_ICUSTAY_INTIME,
  timerange = 24, min_volume_to_include = 0, normalize = FALSE)
}
\arguments{
\item{temp_INS}{A data.frame containing fluids received by the patient}

\item{temp_OUTS}{A data.frame containing fluids out recorded from the patient}

\item{temp_urine}{A data.frame containing urine out recorded from the patient}

\item{temp_ICUSTAY_INTIME}{A subset of ICUSTAY, a data.frame containing
patient identifiers and the time they arrive in the ICU.}

\item{timerange}{A numeric of the number of hours post ICU admission to sum over}

\item{min_volume_to_include}{A numeric of the minimum volume to include in summation}

\item{normalize}{Logical value whether or not to normalize in case length
of stay < timerange}
}
\value{
A list of summation values in, out and urine out.
}
\description{
Input required four data.frames and there are three optional parameters.
}

