% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_oasis_severity_score.R
\name{calculate_oasis_severity_score}
\alias{calculate_oasis_severity_score}
\title{Calculate the OASIS score from the chartevents components.}
\usage{
calculate_oasis_severity_score(OASIS_table)
}
\arguments{
\item{OASIS_table}{A data.frame containing OASIS extreme chartevents for the patient}
}
\value{
A data.frame of the component scores and the final, sum score
}
\description{
Requires a data.frame of all the OASIS component's extreme chartevents.
}

