% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_oasis_severity_score.R
\name{extract_oasis_from_chartevents}
\alias{extract_oasis_from_chartevents}
\title{Extract the OASIS components from chartevents table.}
\usage{
extract_oasis_from_chartevents(CHARTEVENTS, variable, ICUIN)
}
\arguments{
\item{CHARTEVENTS}{A data.frame containing chartevents for the patient}

\item{variable}{A string containing the OASIS parameter to search for.}

\item{ICUIN}{A data.frame subset of the icuin table containing patient
IDs and ICU admission times.}
}
\value{
A vector of the patient ID, minimum, median and maximum variable value
}
\description{
Requires a subset of the chartevents table as a data.frame as well as a
string indicating which variable to look for and the icuin table as a
data.frame to identify the time of ICU admission.
}

