% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_past_medical_history.R
\name{extract_pmhflag_string}
\alias{extract_pmhflag_df}
\alias{extract_pmhflag_df_for_hf}
\alias{extract_pmhflag_string}
\alias{extract_pmhflag_string_for_hf}
\title{Filter a patient based on their admission diagnosis.}
\usage{
extract_pmhflag_string(temp_str, exclusion)

extract_pmhflag_df(temp_df, text_column, exclusion)

extract_pmhflag_string_for_hf(temp_str)

extract_pmhflag_df_for_hf(temp_df, text_column)
}
\arguments{
\item{temp_str}{A string containing the individual's discharge summary note}

\item{exclusion}{A character vector containing the desired exclusion criteria}

\item{temp_df}{A data.frame containing discharge summaries}

\item{text_column}{A character string of the name of the text field}
}
\value{
A vector of logical values describing membership to exclusion
}
\description{
The input can either be a single string or a data.frame in which case the
generic is mapped through each row. The string will be pushed to lowercase.
The string will be searched for past medical history, or history of present
illness. Each substring will be compared against a list of exclusion
critera which if present will be marked as TRUE.
}
\section{Functions}{
\itemize{
\item \code{extract_pmhflag_df}: Apply to data.frame of discharge summaries

\item \code{extract_pmhflag_string_for_hf}: Apply with provided list of heart failure terms

\item \code{extract_pmhflag_df_for_hf}: Apply to data.frame with provided list of heart failure terms
}}
\examples{
# create some examples
temp_str = "blah past medical history: smoking related emphysema. blah: test"
extract_pmh_string(temp_str)
extract_pmhflag_string(temp_str, 'smok') # TRUE
extract_pmhflag_string(temp_str, 'etoh') # FALSE

# data.frame case
temp_df = data.frame(id = c(1,2,3),
  text = c("blah\\nPast medical history: smoking related emphysema. blah: test",
  "blah PRIOR MEDICAL HISTORY: excessive etoh. blah: test",
  "blah history of present illness: obesity, hypertension, HF etc blah: test"))
# mapping
sapply(temp_df$text, function(x) extract_pmhflag_string(x, c('smok', 'etoh')))
# direct
#extract_pmhflag_df(temp_df, 'text', c('smok', 'etoh'))
}

