% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_icd9_diagnosis.R
\name{filter_icd9_by_codes}
\alias{filter_icd9_by_codes}
\alias{filter_icd9_by_severesepsis}
\title{Filter a set of ICD-9 codes against a reference set.}
\usage{
filter_icd9_by_codes(test_codes, reference_codes)

filter_icd9_by_severesepsis(test_codes)
}
\arguments{
\item{test_codes}{A vector or list ofICD-9 value to test.}

\item{reference_codes}{A vector or list of ICD-9 codes to act as reference.}
}
\value{
A vector of logical (TRUE/FALSE) values representing inclusion/exclusion.
}
\description{
Inputs can be either character or numeric lists or vectors but are coerced
to character vectors. The \code{test_codes} argument is assumed to be
completely specified - that is, two values after the '.' period. The
\code{icd9.reference} set does not have to be completely specified so that,
'123' includes '123.34' and '123.45'.
}
\section{Functions}{
\itemize{
\item \code{filter_icd9_by_severesepsis}: Apply filtering with two set lists for severe sepsis
}}
\examples{
# With dplyr
data = data.frame(id = c(1, 2, 3, 4, 5, 6),
  ICD9 = c('123.45', '234.56', '345.67', '456.78','567.89', '678.90'))
dplyr::filter(data, filter_icd9_by_codes(ICD9, c('123', '345', '567')))

# Without dplyr
mask = filter_icd9_by_codes(data$ICD9, c('123', '345', '567'))
data.filtered = data[mask,]

}

