% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sepsis_admdx_exclusion}
\alias{sepsis_admdx_exclusion}
\title{List of admission diagnoses to exclude for a severe sepsis cohort.}
\format{\describe{
  \item{sepsis_admdx_exclusion}{list of exclusion criter}
}}
\source{
Monique Tanna
}
\usage{
sepsis_admdx_exclusion
}
\description{
A list of admission diagnoses that do not indicate a manifestation of
severe sepsis and should be excluded from the filtered cohort.
}
\keyword{datasets}

