% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sepsis_admdx_inclusion}
\alias{sepsis_admdx_inclusion}
\title{List of admission diagnoses to include for a severe sepsis cohort.}
\format{\describe{
  \item{sepsis_admdx_inclusion}{list of inclusion criter}
}}
\source{
Monique Tanna
}
\usage{
sepsis_admdx_inclusion
}
\description{
A list of admission diagnoses that do indicate a manifestation of
severe sepsis and should be included in the filtered cohort.
}
\keyword{datasets}

