% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sepsis_icd9_infection}
\alias{sepsis_icd9_infection}
\title{Data.frame of ICD-9 diagnosis codes for systemic infection and their meanings.}
\format{\describe{
  \item{sepsis_icd9_infection}{data.frame of systemic infection ICD-9 codes and their meanings}
}}
\source{
Monique Tanna
}
\usage{
sepsis_icd9_infection
}
\description{
A list of ICD-9 diagnosis codes that indicate some kind of systemic
infection - one essential part of severe sepsis.
}
\keyword{datasets}

