% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sepsis_icd9_organfailure}
\alias{sepsis_icd9_organfailure}
\title{Data.frame of ICD-9 diagnosis codes for organ failure and their meanings.}
\format{\describe{
  \item{sepsis_icd9_organfailure}{data.frame of organ failure ICD-9 codes and their meanings}
}}
\source{
Monique Tanna
}
\usage{
sepsis_icd9_organfailure
}
\description{
A list of ICD-9 diagnosis codes that indicate some kind of major organ
failure - one essential part of severe sepsis.
}
\keyword{datasets}

